refactor store header menu component
