+++ src/app/store/header/menu/store-header-menu.component.html
@@ -1,0 +1,27 @@
+<mat-menu #headerMenu>
+  <span mat-menu-item disabled>{{ userName$ | async }}</span>
+  <button mat-menu-item type="button"
+    *ngIf="canNavigateManagement$ | async"
+    class="edit-profile"
+    [routerLink]="['/management']">
+    <mat-icon>tune</mat-icon>
+    Administrar
+  </button>
+  <button mat-menu-item type="button"
+    class="edit-profile"
+    (click)="onClickEditProfile()">
+    <mat-icon>edit</mat-icon>
+    Editar Perfil
+  </button>
+  <button mat-menu-item type="button"
+    class="logout"
+    (click)="onClickLogout()">
+    <mat-icon>exit_to_app</mat-icon>
+    Terminar Sesión
+  </button>
+</mat-menu>
+<button mat-button type="button"
+  class="open-menu" title="Menú de usuario"
+  [matMenuTriggerFor]="headerMenu">
+  <mat-icon>person</mat-icon>
+</button>
