+++ src/app/store/header/menu/store-header-menu.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { StoreHeaderMenuComponent } from './store-header-menu.component';
+
+describe('StoreHeaderMenuComponent', () => {
+  let component: StoreHeaderMenuComponent;
+  let fixture: ComponentFixture<StoreHeaderMenuComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ StoreHeaderMenuComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(StoreHeaderMenuComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
