+++ src/app/store/header/menu/store-header-menu.component.ts
@@ -1,0 +1,88 @@
+import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { Observable, of } from 'rxjs';
+import { pluck, switchMap } from 'rxjs/operators';
+import { AppService } from 'src/app/app.service';
+import { ConfirmationDialogComponent, ConfirmationDialogData } from 'src/app/shared/confirmation-dialog/confirmation-dialog.component';
+import { EditProfileFormDialogComponent } from 'src/app/shared/edit-profile-form-dialog/edit-profile-form-dialog.component';
+import { LOGOUT_MESSAGE } from 'src/text/messages';
+
+@Component({
+  selector: 'app-store-header-menu',
+  templateUrl: './store-header-menu.component.html',
+  styleUrls: ['./store-header-menu.component.css']
+})
+export class StoreHeaderMenuComponent
+  implements OnInit {
+
+  public userName$: Observable<string>;
+  public canNavigateManagement$: Observable<boolean>;
+
+  constructor(
+    protected appService: AppService,
+    protected snackBarService: MatSnackBar,
+    protected dialogService: MatDialog
+  ) { }
+
+  ngOnInit(): void {
+
+    this.userName$ = this.appService.isLoggedInChanges$.pipe(
+      switchMap(
+        (isLoggedIn: boolean) => {
+          if (!isLoggedIn) {
+            return of('');
+          } else {
+            return this.appService.getUserProfile().pipe(pluck('name'));
+          }
+        }
+      )
+    );
+
+    this.canNavigateManagement$ = this.appService.isLoggedInChanges$.pipe(
+      switchMap(
+        (isLoggedIn: boolean) => {
+          return of(true);
+        }
+      )
+    );
+  }
+
+  protected promptLogoutConfirmation(): Observable<boolean> {
+    const dialogData: ConfirmationDialogData = {
+      title: '¿Cerrar sesion?',
+      message: 'Si esta realizando una transaccion, perdera la informacion que haya guardado.'
+    };
+
+    return this.dialogService.open(
+      ConfirmationDialogComponent,
+      {
+        width: '24rem',
+        data: dialogData
+      }
+    ).afterClosed();
+  }
+
+  public onClickEditProfile(): void {
+    this.dialogService.open(
+      EditProfileFormDialogComponent,
+      {
+        width: '60rem'
+      }
+    );
+  }
+
+  public onClickLogout(): void {
+    if (this.appService.isLoggedIn()) {
+      this.promptLogoutConfirmation().subscribe(
+        confirmed => {
+          if (confirmed) {
+            this.snackBarService.open(LOGOUT_MESSAGE, 'OK');
+            this.appService.closeCurrentSession();
+          }
+        }
+      );
+    }
+  }
+
+}
