+++ src/app/store/header/store-header.component.css
@@ -30,6 +30,10 @@ app-store-header-middle {
   flex: 1 1 auto;
 }
 
+app-store-header-menu {
+  flex: 0 0 auto;
+}
+
 .action {
   flex: 0 0 auto;
   justify-content: center;
