+++ src/app/store/header/store-header.component.html
@@ -10,33 +10,7 @@
   <app-store-header-navigation></app-store-header-navigation>
   <app-store-header-middle></app-store-header-middle>
   <ng-container *ngIf="(isLoggedIn$ | async); else notLoggedIn">
-    <mat-menu #headerMenu>
-      <span mat-menu-item disabled>{{ userName$ | async }}</span>
-      <button mat-menu-item type="button"
-        *ngIf="canNavigateManagement$ | async"
-        class="edit-profile"
-        [routerLink]="['/management']">
-        <mat-icon>tune</mat-icon>
-        Administrar
-      </button>
-      <button mat-menu-item type="button"
-        class="edit-profile"
-        (click)="onClickEditProfile()">
-        <mat-icon>edit</mat-icon>
-        Editar Perfil
-      </button>
-      <button mat-menu-item type="button"
-        class="logout"
-        (click)="onClickLogout()">
-        <mat-icon>exit_to_app</mat-icon>
-        Terminar Sesión
-      </button>
-    </mat-menu>
-    <button mat-button type="button"
-      class="open-menu action" title="Menú de usuario"
-      [matMenuTriggerFor]="headerMenu">
-      <mat-icon>person</mat-icon>
-    </button>
+    <app-store-header-menu></app-store-header-menu>
   </ng-container>
   <ng-template #notLoggedIn>
     <button mat-button type="button"
