+++ src/app/store/header/store-header.component.ts
@@ -5,14 +5,11 @@
 
 import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
 import { Observable, of } from 'rxjs';
-import { pluck, switchMap, startWith } from 'rxjs/operators';
+import { startWith } from 'rxjs/operators';
 import { AppService } from 'src/app/app.service';
 import { ConfirmationDialogComponent, ConfirmationDialogData } from 'src/app/shared/confirmation-dialog/confirmation-dialog.component';
-import { EditProfileFormDialogComponent } from 'src/app/shared/edit-profile-form-dialog/edit-profile-form-dialog.component';
-import { LOGOUT_MESSAGE } from 'src/text/messages';
 import { StoreLoginFormDialogComponent } from '../dialogs/login-form/store-login-form-dialog.component';
 
 @Component({
@@ -25,12 +22,8 @@ export class StoreHeaderComponent
 
   public isLoggedIn$: Observable<boolean>;
 
-  public userName$: Observable<string>;
-  public canNavigateManagement$: Observable<boolean>;
-
   constructor(
     protected appService: AppService,
-    protected snackBarService: MatSnackBar,
     protected dialogService: MatDialog,
     protected router: Router
   ) { }
@@ -40,41 +33,6 @@ export class StoreHeaderComponent
     this.isLoggedIn$ = this.appService.isLoggedInChanges$.pipe(
       startWith(this.appService.isLoggedIn())
     );
-
-    this.userName$ = this.appService.isLoggedInChanges$.pipe(
-      switchMap(
-        (isLoggedIn: boolean) => {
-          if (!isLoggedIn) {
-            return of('');
-          } else {
-            return this.appService.getUserProfile().pipe(pluck('name'));
-          }
-        }
-      )
-    );
-
-    this.canNavigateManagement$ = this.appService.isLoggedInChanges$.pipe(
-      switchMap(
-        (isLoggedIn: boolean) => {
-          return of(true);
-        }
-      )
-    );
-  }
-
-  protected promptLogoutConfirmation(): Observable<boolean> {
-    const dialogData: ConfirmationDialogData = {
-      title: '¿Cerrar sesion?',
-      message: 'Si esta realizando una transaccion, perdera la informacion que haya guardado.'
-    };
-
-    return this.dialogService.open(
-      ConfirmationDialogComponent,
-      {
-        width: '24rem',
-        data: dialogData
-      }
-    ).afterClosed();
   }
 
   protected promptManagementRedirect(): void {
@@ -104,28 +62,6 @@ export class StoreHeaderComponent
         width: '24rem'
       }
     );
-  }
-
-  public onClickEditProfile(): void {
-    this.dialogService.open(
-      EditProfileFormDialogComponent,
-      {
-        width: '60rem'
-      }
-    );
-  }
-
-  public onClickLogout(): void {
-    if (this.appService.isLoggedIn()) {
-      this.promptLogoutConfirmation().subscribe(
-        confirmed => {
-          if (confirmed) {
-            this.snackBarService.open(LOGOUT_MESSAGE, 'OK');
-            this.appService.closeCurrentSession();
-          }
-        }
-      );
-    }
   }
 
 }
